services/clientreport/clientapp/serializers.py
from rest_framework import serializers
from .models import Client, Address

class AddressSerializer(serializers.ModelSerializer):
    class Meta:
        model = Address
        fields = ['id', 'description']

class ClientSerializer(serializers.ModelSerializer):
    addresses = AddressSerializer(many=True, read_only=True)

    class Meta:
        model = Client
        fields = ['id', 'name', 'addresses']

