services/clientreport/requirements.txt
psycopg2-binary==2.9.6
asgiref==3.8.1
confluent-kafka==2.6.1
Django==5.1.4
djangorestframework==3.15.2
sqlparse==0.5.3
